rectangle area through diagonal and side area using math fcts

+++ day03/ex03.cpp
@@ -4,7 +4,7 @@
 int main(void)
 {
     int a, d;
-    int Area;
+    float Area;
 
     std::cout << "Please enter a : ";
     std::cin >> a;
